+++ g2g-plugins/src/lib.rs
@@ -10,6 +10,7 @@
 extern crate alloc;
 
 pub mod aacparse;
+pub mod opusparse;
 pub mod audioconvert;
 pub mod audioresample;
 pub mod audiotestsrc;